requirements.txt
Flask
APScheduler
requests
mailjet-rest
premailer
selenium
jinja2
google-generativeai
